Fix a bug that the type of ``nX`` in ``bisection_search`` is wrong. Change ``int`` type varialbes in ``bisection_search`` and ``interp`` to ``unsigned int``.

+++ mag_calc_cext.c
@@ -140,13 +140,13 @@ void free_2d_double(double **p, int nRow) {
 }
 
 
-inline int bisection_search(double a, double *x, double nX) {
+inline unsigned int bisection_search(double a, double *x, int nX) {
     /* return idx such x[idx] <= a < x[idx + 1] 
      * a must be x[0] <= a < x[nX - 1]
      */
-    int idx0 = 0;
-    int idx1 = nX - 1;
-    int idxMid;
+    unsigned int idx0 = 0;
+    unsigned int idx1 = nX - 1;
+    unsigned int idxMid;
     while(idx1 - idx0 > 1) {
         idxMid = (idx0 + idx1)/2;
         if(a >= x[idxMid])
@@ -160,7 +160,7 @@ inline int bisection_search(double a, double *x, double nX) {
 
 inline double interp(double xp, double *x, double *y, int nPts) {
     /* Interpolate a given points */
-    int idx0, idx1;
+    unsigned int idx0, idx1;
     if((xp < x[0]) || (xp > x[nPts - 1])) {
         printf("Error: Point %10.5e is beyond the interpolation region\n", xp);
         exit(0);
